requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-braces==1.17.0
django-debug-toolbar==5.2.0
django-embed-video==1.4.10
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
pymemcache==4.0.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
